add redirection for nohup and stop server.sh

+++ scripts/stop_server.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 
-# kills all processes PID running on port 80 - where flask server is running
 pid=$(sudo lsof -ti TCP:80)
-echo $pid
-sudo kill -9 $pid
+
+if [ -z "$pid" ]; then
+    echo "no processes running on port 80"
+else
+    echo "killing process ${pid} running on port 80"
+    sudo kill -9 $pid
+fi
+
